move ts completion to new file, also move getLSAndTSDoc to new class
